Simplified framework for inversion
