+++ simpegtorch/utils/topography_utils.py
@@ -186,105 +186,6 @@ def active_from_xyz(
     )
 
 
-def drape_electrodes_on_topography(
-    survey, surface_xyz: Union[torch.Tensor, np.ndarray], method="linear"
-):
-    """
-    Drape electrode locations onto flat topography.
-
-    This is a simplified implementation for flat surfaces.
-
-    Parameters
-    ----------
-    survey : Survey
-        The survey object containing electrode locations
-    surface_xyz : torch.Tensor or numpy.ndarray
-        Surface points defining the topography
-    method : str, default="linear"
-        Interpolation method (only "linear" supported)
-
-    Returns
-    -------
-    Survey
-        Survey object with updated electrode locations
-
-    Examples
-    --------
-    >>> # For flat topography, this function updates electrode z-coordinates
-    >>> # to match the surface elevation
-    >>> draped_survey = drape_electrodes_on_topography(survey, surface_xyz)
-    """
-
-    # Convert inputs to torch tensors
-    if not isinstance(surface_xyz, torch.Tensor):
-        surface_xyz = torch.tensor(surface_xyz, dtype=torch.float64)
-
-    # For flat topography, use mean surface elevation
-    surface_z = torch.mean(surface_xyz[:, 2])
-
-    print(f"Draping electrodes to surface elevation: {surface_z:.2f} m")
-
-    # Update electrode locations for all sources
-    for src in survey.source_list:
-        # Update source location
-        if hasattr(src, "location") and src.location is not None:
-            # Convert to tensor if needed and extract scalar values
-            if isinstance(src.location, torch.Tensor):
-                # Handle different tensor shapes
-                if src.location.numel() == 1:
-                    # Single value tensor
-                    x_val = src.location.item()
-                    y_val = 0.0  # Default
-                elif src.location.numel() >= 2:
-                    # Multi-element tensor
-                    x_val = (
-                        src.location[0].item()
-                        if src.location[0].numel() == 1
-                        else src.location[0]
-                    )
-                    y_val = (
-                        src.location[1].item()
-                        if src.location[1].numel() == 1
-                        else src.location[1]
-                    )
-                else:
-                    continue  # Skip if invalid
-            else:
-                x_val = src.location[0]
-                y_val = src.location[1]
-
-            # Ensure we have scalar values
-            if isinstance(x_val, torch.Tensor):
-                x_val = x_val.item()
-            if isinstance(y_val, torch.Tensor):
-                y_val = y_val.item()
-
-            src.location = torch.tensor(
-                [x_val, y_val, surface_z.item()], dtype=torch.float64
-            )
-
-        # Update receiver locations
-        for rx in src.receiver_list:
-            if hasattr(rx, "locations") and rx.locations is not None:
-                # Ensure rx.locations is a tensor
-                if not isinstance(rx.locations, torch.Tensor):
-                    rx.locations = torch.tensor(rx.locations, dtype=torch.float64)
-
-                rx.locations = torch.column_stack(
-                    [
-                        rx.locations[:, 0],  # x
-                        rx.locations[:, 1],  # y
-                        torch.full(
-                            (rx.locations.shape[0],),
-                            surface_z.item(),
-                            dtype=torch.float64,
-                        ),  # z
-                    ]
-                )
-
-    return survey
-
-
 def create_flat_topography(
     x_extent, y_extent, elevation=0.0, n_points_x=51, n_points_y=51
 ):
